Rename CSS smoke test to language features: For #90538
