+++ test/smoke/src/areas/languages/languages.test.ts
@@ -3,10 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { Application, ProblemSeverity, Problems } from '../../../../automation';
+import { Application, ProblemSeverity, Problems } from '../../../../automation/out';
 
 export function setup() {
-	describe('Languages - CSS', () => {
+	describe('Language Features', () => {
 		it('verifies quick outline', async function () {
 			const app = this.app as Application;
 			await app.workbench.quickopen.openFile('style.css');
@@ -15,7 +15,7 @@ export function setup() {
 			await app.workbench.quickopen.waitForQuickOpenElements(names => names.length === 2);
 		});
 
-		it('verifies warnings for the empty rule', async function () {
+		it('verifies problems view', async function () {
 			const app = this.app as Application;
 			await app.workbench.quickopen.openFile('style.css');
 			await app.workbench.editor.waitForTypeInEditor('style.css', '.foo{}');
@@ -27,7 +27,7 @@ export function setup() {
 			await app.workbench.problems.hideProblemsView();
 		});
 
-		it('verifies that warning becomes an error once setting changed', async function () {
+		it('verifies settings', async function () {
 			const app = this.app as Application;
 			await app.workbench.settingsEditor.addUserSetting('css.lint.emptyRules', '"error"');
 			await app.workbench.quickopen.openFile('style.css');
